requirements.txt
feedparser
eyed3
python-dateutil
